requirements.txt
streamlit
langchain
langchain-community
langchain-groq
serpapi
python-dotenv
google-search-results
